Moving into the required directory structure for the includes. Adding an xcode project to build a static cocoa touch library.

+++ Box2D/Dynamics/Controllers/b2GravityController.cpp
@@ -16,7 +16,7 @@
 * 3. This notice may not be removed or altered from any source distribution.
 */
 
-#include "b2GravityController.h"
+#include <Box2D/Dynamics/Controllers/b2GravityController.h>
 
 b2GravityController::b2GravityController(const b2GravityControllerDef* def) : b2Controller(def)
 {
@@ -36,7 +36,7 @@ void b2GravityController::Step(const b2TimeStep& step)
 				float32 r2 = d.LengthSquared();
 				if(r2 < b2_epsilon)
 					continue;
-				b2Vec2 f = G / r2 / sqrt(r2) * body1->GetMass() * body2->GetMass() * d;
+				b2Vec2 f = G / r2 / sqrtf(r2) * body1->GetMass() * body2->GetMass() * d;
 				body1->ApplyForce(f      , body1->GetWorldCenter());
 				body2->ApplyForce(-1.0f*f, body2->GetWorldCenter());
 			}
